requirements.txt
pdfplumber==0.11.4
spacy==3.7.6
sentence-transformers==2.7.0 
faiss-cpu==1.8.0
rank-bm25==0.2.2
transformers==4.44.2
torch==2.4.1  
flask==3.0.3
gunicorn==23.0.0
numpy==1.26.4
pyyaml==6.0.2
python-dotenv==1.0.1
google-generativeai==0.7.2